+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 
 ## [Unreleased]
 
+## [0.1.2] - 2021.06.28
+### Added
+- Database Editor Window에서 treeMenu를 더블클릭하면 프로젝트 뷰에서 해당 에셋을 인스펙터로 선택하는 기능 추가.
+
+
+## [0.1.1] - 2021.06.28
+### Added
+- IEquipment 인터페이스를 WeaponItem 및 ArmorItem에 구현
+- ItemBase에서 GetCopy로 아이템 사본을 얻는 기능 추가.
+
 ## [0.1.0] - 2021.06.28
 ### Added
 - 간단한 인벤토리 시스템(Inventory) 추가